backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "arxiv>=2.3.1",
    "fastapi>=0.124.0",
    "google-genai>=1.54.0",
    "langchain-google-genai>=4.0.0",
    "langgraph>=1.0.4",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "psycopg[async,binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncpg]>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
